requirements.txt
Django==4.1
django-cors-headers==3.13.0
django-cors-middleware==1.5.0
django-environ==0.9.0
django-oauth-toolkit==2.2.0
djangorestframework==3.14.0
drf-access-policy==1.3.0
drf-yasg==1.21.4
gunicorn==20.1.0
oauthlib==3.2.2
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
python-decouple==3.6
pytz==2022.6
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
drf-standardized-errors==0.12.3
whitenoise